backend/requirements.txt
openpyxl
xlrd
nltk
pystardict
pdfplumber==0.9.0
pypdf2
bs4
Django
django_bootstrap5
python_docx==1.0.1
baidu-aip
PyMuPDF==1.23.3
markdownify==0.11.6
mobi==0.3.3
ebooklib
mistune==3.0.2
langdetect==1.0.9
iso639
PyYAML==6.0.1
pandas==2.1.3
chardet==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
coreapi==2.3.3
psycopg==3.1.13
html2text==2020.1.16
Levenshtein==0.23.0
python-dotenv==1.0.0

matplotlib==3.8.2
seaborn==0.13.1
jieba==0.42.1
wordcloud==1.9.3
python-frontmatter==1.0.1
#websocket==0.2.1
#websocket-client==0.56.0
websocket-client-py3==0.15.0

pydub==0.25.1
easyliter==1.0.5
gTTS==2.5.0

librosa==0.10.1
minio==7.2.5

langchain==0.1.14
langchain-google-genai==1.0.1
langchain-openai==0.1.5
google-generativeai==0.4.1
openai==1.18.0
tiktoken==0.7.0
pyautogen==0.2.35

django==5.0.6
django-bootstrap5==24.2
django-cors-headers==4.3.1
django-cron==0.6.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0

edge_tts
pgvector

